+++ include/libdpf/libdpf.h
@@ -5,8 +5,8 @@
 #include "block.h"
 #include "typedef.h"
 
-int GEN(AES_KEY *key, uint64_t alpha, const uint n, uchar **k0, uchar **k1);
-uint128 EVAL(AES_KEY *key, uchar *k, uint64_t x);
+int GEN(AES_KEY *key, uint alpha, const uint n, uchar **k0, uchar **k1);
+uint128 EVAL(AES_KEY *key, uchar *k, uint x);
 uint128 *EVALFULL(AES_KEY *key, const uchar *k);
 
 void test_libdpf();
